In case we do replace Fragment Issue Added. @Override public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) { if (rootView == null) { } return rootView; }
In case of if we do replace fragment the beneath fragment gets replaced with new one and later on back press it gets recreated. If Beneath fragment had a request call in onCreateView then this issue would have arised i.e OnCreateView would have been called again and also the Request call would have been made as the fragment was replaced and it gets recreated when we backpress.
We can solve this issue : - Since view of Fragment is saved and is not destroyed we can reuse the same view by writing the above code.